This is the finished project
For the students who will use this because I know you will. Actually use this to learn please. You will be tested on this and the work in this lab is crucial for further development in computer science. If/when you copy stuff learn about it, use this as a helper; an accelerator in this project, not a crutch so you can pull an easy A and not learn anything.